Refactor : 컴포넌트 명칭 통일화 (store -> restaurant) (#65)
