+++ frontend/src/components/Mypage/RestaurantManage/RestaurantList.tsx
@@ -1,10 +1,10 @@
 import { useState } from "react";
-import type { StoreFormRequest } from "../../../types/storeFormRequest";
+import type { StoreFormRequest } from "../../../types/restaurantFormRequest.ts";
 import { dummyRestaurantDetails } from "../../../data/dummyRestaurantDetail";
 import { FiChevronLeft, FiEdit2, FiPlus, FiTrash2 } from "react-icons/fi";
 import DetailStores from "../../RestaurantDetail/RestaurantDetail.tsx";
 import type { RestaurantDetail } from "../../../types/restaurant.types";
-import StoreFormModal from "./modals/StoreFormModal";
+import StoreFormModal from "./modals/ResaurantFormModal.tsx";
 import Button from "../../Common/Button";
 
 // 변환 함수: RestaurantDetail → StoreFormRequest
